==> Lessons/11_Lab_CraftingReports.Rmd <==
---
title: 'Lab 11: Crafting Reports'
author: Environmental Data Analytics | John Fay & Luana Lima | Developed by Kateri Salk
date: "Spring 2021"
output:
 pdf_document: default

---

## OBJECTIVES

1. Questions on M10/A10
2. Share updated syllabus
3. Talk about final project (groups/prompt/rubric)
    Enter groups [here][https://docs.google.com/spreadsheets/d/16LHgSIx3SLQNjyWSKMkFvPCc26phm38q5F3crwZIfGY/edit?usp=sharing] 
4. Questions on M11
5. Exercise 1: Fill lesson file for M11
6. Exercise 2: Shiny app
